// src/app/shared/models/visita.model.ts
// Generated by https://quicktype.io

export interface Visita {
    id:                 number;
    correlativo?:      string;
    idEmpresa:          number;
    idUsuario?:          number;
    usuario?:      string;
    idInteresado?:      number;
    interesado?:      string;
    direccion?:      string;
    idEstadoVisita: number;
    estadoVisita:        string;
    fecha?:              string;
    idMotivoCancelacion?: number;
    motivoCancelacion?:        string;
    inicioVisita?:              Date;
    finVisita?:              Date;
    duracionVisita?:        number;
    detalle?:            VisitaDetalle;
    latitud?:  number;
	longitud?: number;

}

export interface VisitaDetalle {
    id?:              number;
    idVisita?:         number;
    detalle?:      string;
    observaciones?:     string;
    imaganes?:     string;
}

export interface VisitaRegistro {
    idEmpresa:          number;
    listaAsignacion:  DireccionUsuario[];
    fecha:              string;
}

export interface DireccionUsuario {
    idInteresadoDireccion: number;
    idUsuario: number;
}

export interface VisitaPosicion {
    latitud?:  number;
	longitud?: number;
    interesado?:      string;
    idEstadoVisita: number;
    estadoVisita:        string;
}